feat: updating aws-* for features and reduced complexity
The major update here is the change of algorithm for sms region selection: https://github.com/tradle/aws-utils/commit/1deb0d71036e5159089cba975528b828c9fd1dbc

The .idx is now part of the aws-info. We can access it through @aws-common-utils in order to make sure that the same data-source is used. This makes the code a bit lighter as well.