feat: Adding support and documentation for dynamic plugins.
This includes documentation in ./docs/plugins.md on the way the plugins work.

Note:
I am not happy about this implementation. It is done in order to have a working system but down-the-line I am reasonably certain that this will cause issues.

Currently the dynamic plugins are processed basically separately from the regular plugins and that makes a lot of things difficult. The, imo, correct alternative
to this implementation is a complex refactoring that loads all plugins using the same system. However, that may have a performance impact if done poorly.

Due to time-constraints I am going with this code that contains code duplication. I am sorry - orz.